Separate pattern into its own parameter for ack-and-a-half-run
Originally the pattern to search for was part of the extra "arguments"
parameter which didn't play nice with the change to be able to search
for patterns that began with a dash.  This came up, specifically, when
trying to use ack-and-a-half-same, since arguments would be something
like "--type elisp 'ack-and-a-half'", which would be placed after the
end-of-arguments marker.

We now pass the pattern to search for separately from additional
arguments to be able to reliably combine everything in the proper
order, with additional arguments before the end-of-arguments marker,
and the pattern after it.

Thanks to Diep Pham Van for noticing this.