Do not claim ack exited abnormally when no matches were found
While ack does exit non-zero when there are no matches found, this is not
actually exiting "abnormally".  Instead, lets behave a bit closer to what
grep.el does, and actually look at what ack did to see if we should be
claiming that it exited abnormally, or adjusting the message to inform the
user that there weren't any matches.  Addresses issue #20.